# Speed-up matrix construction by ensuring MatrixArgs type MA_ENTRIES_ZERO